Zero-copy astype

+++ python/cudf/cudf/core/dataframe.py
@@ -5435,10 +5435,10 @@ class DataFrame(Frame, Serializable):
         else:
             mask = None
 
-        coerced = filtered.astype(common_dtype, copy=True)
+        coerced = filtered.astype(common_dtype)
         if is_pure_dt:
             # Further convert into cupy friendly types
-            coerced = coerced.astype("int64", copy=False)
+            coerced = coerced.astype("int64")
         return coerced, mask, common_dtype
 
     def count(self, axis=0, level=None, numeric_only=False, **kwargs):
